добавил функцию ручного добавления сотрудника

+++ Lab1New/Input.h
@@ -3,6 +3,6 @@
 
 int InputInt(std::string message, int min, int max);
 
-double InputDouble(std::string message, int min, int max);
+double InputDouble(std::string message, double min, double max);
 
 std::string InputString(std::string message);
